+++ objectbox/CHANGELOG.md
@@ -11,6 +11,8 @@
 * Generator not longer warns that it can not find the package source root if the output directory is
   the package root directory.
 * Add `StorageException` which is a `ObjectBoxException` with an `errorCode` (a `OBX_ERROR` code).
+* Throw `DbFullException` instead of `ObjectBoxException` with message `10101 Could not put` (error 
+  code `OBX_ERROR_DB_FULL`).
 
 ## 1.6.2 (2022-08-24)